+++ src/test/java/seedu/duke/TestManagerTest.java
@@ -1,23 +1,4 @@
 package seedu.duke;
 
-import org.junit.jupiter.api.Test;
-import seedu.duke.exceptions.FieldEmptyException;
-import seedu.duke.parser.TestParser;
-import seedu.duke.testing.TestManager;
-
-import static org.junit.jupiter.api.Assertions.assertThrows;
-
 public class TestManagerTest {
-
-    @Test
-    public void parseUserResponse_noInput_expectFieldEmptyException() {
-        String input = "";
-        assertThrows(FieldEmptyException.class, () -> TestParser.parseUserResponse(input));
-    }
-
-    @Test
-    public void toInt_invalidInput_expectNumberFormatException() {
-        String input = "a";
-        assertThrows(NumberFormatException.class, () -> TestParser.toInt(input));
-    }
 }
